project/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
langgraph>=0.0.20
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
plotly>=5.18.0
yfinance>=0.2.0
feedparser>=6.0.10
